requirements.txt
aioredis==1.3.1
amqp==5.1.1
APScheduler==3.10.4
arrow==1.2.3
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.1.0
blessed==1.20.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==15.1.0
cryptography==41.0.4
daphne==3.0.2
Django==4.0
django-apscheduler==0.6.2
django-cors-headers==3.13.0
django-db-signals==0.1.1
django-filter==22.1
django-picklefield==3.1
django-q==1.3.9
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
hiredis==2.2.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.3.2
msgpack==1.0.6
prompt-toolkit==3.0.39
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==23.2.0
pyst2==0.5.1
python-dateutil==2.8.2
pytz==2023.3.post1
redis==3.5.3
requests==2.31.0
schedule==1.2.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.6
zope.interface==6.0
